move fn to p4util.text
