+++ psi4/driver/p4util/text.py
@@ -226,3 +226,38 @@ def find_approximate_string_matches(seq1, options, max_distance):
     """Find list of approximate (within `max_distance`) matches to string `seq1` among `options`."""
 
     return [seq2 for seq2 in options if (levenshtein(seq1, seq2) <= max_distance)]
+
+def message_box(message: str = None, max_width: int = 80):
+    """ put a message string into a box for extra attention
+
+    Parameters
+    -----------
+    message
+       message string to be boxed
+
+    max_width
+        maximal character width of the box
+
+    Returns
+    --------
+    str
+       box containing the message as a multiline string
+    """
+    from textwrap import wrap
+    
+    is_long = (len(message) > max_width)
+    box_width = max_width if is_long else len(message)
+
+    message = " ".join(message.split("\n"))
+    message = wrap(message, box_width, break_long_words=False)
+    error_str = []
+    error_str.append('\n!' + '-' * box_width + '--!\n')
+
+    fmt = "! {:" + str(box_width) + "} !\n"
+    for line in message:
+        error_str.append(fmt.format(line))
+
+    error_str.append('!' + '-' * box_width + '--!\n')
+    error_str = ''.join(error_str)
+
+    return error_str
