+++ psi4/run_psi4.py
@@ -357,9 +357,9 @@ except Exception as exception:
     ex = ','.join(traceback.format_exception_only(type(exception), exception))
     ex_list = ex.split(":", 1)[-1]
     error = ''.join(ex_list)
-    psi4.core.print_out(psi4.extras.message_box(error))
+    psi4.core.print_out(psi4.driver.p4util.text.message_box(error))
     if psi4.core.get_output_file() != "stdout":
         print(tb_str)
         print(in_str)
-        print(psi4.extras.message_box(error))
+        print(psi4.driver.p4util.text.message_box(error))
     sys.exit(1)
